Server validation of add form and notification of success
- Server validation
- If not valid, keep old data
- Redirects to Map when valid
- Displays success notification
- CSS corrections

+++ SaltedButterWebsite/Models/NokiaPlace.cs
@@ -16,7 +16,7 @@ namespace SaltedButterWebsite.Models
 
     public class Location
     {
-        public decimal[] Position { get; set; }
+        public double[] Position { get; set; }
         public Dictionary<string,string> Address { get; set; }
         //public decimal[] Bbox { get; set; }
